refactor(actions): tags/ como directorio con barril
